+++ src/datumaro/plugins/specs.py
@@ -56,4 +56,9 @@ if __name__ == "__main__":
     _enroll_to_plugin_specs(env.transforms, "Transform")
     _enroll_to_plugin_specs(env.validators, "Validator")
 
-    dump_json_file(_SPECS_JSON_PATH, plugin_specs, indent=True, append_newline=True)
+    dump_json_file(
+        _SPECS_JSON_PATH,
+        sorted(plugin_specs, key=lambda spec: spec["import_path"]),
+        indent=True,
+        append_newline=True,
+    )